Reorganize files into qa_utils dir
